Rename apiType to wireApi and split adapters into providers/
- Config/settings field is wireApi (WireApi, DEFAULT_WIRE_API, wireApiFromString).
- Default remains chat_completions when wireApi is omitted.
- Move adapters under aiclient/src/providers/:
  types, shared, chatCompletions, openaiResponses, anthropicMessages,
  providerAdapter dispatch, index barrel.
