+++ ts/config.sample.yaml
@@ -86,7 +86,7 @@ azureOpenAI:
     gpt_v:
       - endpoint: https://my-resource.openai.azure.com/openai/deployments/gpt-v/chat/completions?api-version=2025-01-01-preview
 
-    # `apiType` selects the wire protocol an endpoint speaks. It is
+    # `wireApi` selects the wire protocol an endpoint speaks. It is
     # applied *after* the routing pool has picked an endpoint, so it does
     # not affect priority/capacity/cooldown routing at all. Omit it (as
     # every deployment above does) to get the default `chat_completions`
@@ -97,13 +97,13 @@ azureOpenAI:
       endpoints:
         - endpoint: https://api.openai.com/v1/responses
           region: eastus
-          apiType: openai_responses
+          wireApi: openai_responses
 
     claude_sonnet:
       endpoints:
         - endpoint: https://api.anthropic.com/v1/messages
           region: eastus
-          apiType: anthropic_messages
+          wireApi: anthropic_messages
 
 # ╔═══════════════════════════════════════════════════════════════════╗
 # ║  OpenAI (non-Azure)                                              ║
