+++ ts/packages/config/README.md
@@ -41,52 +41,52 @@ YAML maps are flattened into the `EnvVars` flat-key shape used by
 
 See [src/flatten.ts](./src/flatten.ts) for the full mapping.
 
-## Multi-provider `apiType`
+## Multi-provider `wireApi`
 
-An Azure OpenAI deployment endpoint may declare an `apiType` selecting the
+An Azure OpenAI deployment endpoint may declare a `wireApi` selecting the
 **wire protocol** it speaks. This lets a single TypeAgent config route to
 Azure/OpenAI Chat Completions, the OpenAI Responses API, and the
 Anthropic Messages API behind the same endpoint pool.
 
-| `apiType` value      | Wire protocol                             | Default? |
+| `wireApi` value      | Wire protocol                             | Default? |
 | -------------------- | ----------------------------------------- | -------- |
 | `chat_completions`   | OpenAI / Azure OpenAI `/chat/completions` | ✅ yes   |
 | `openai_responses`   | OpenAI `/responses`                       | no       |
 | `anthropic_messages` | Anthropic `/v1/messages`                  | no       |
 
-**Default & back-compat.** `apiType` is optional. Omitting it is exactly
+**Default & back-compat.** `wireApi` is optional. Omitting it is exactly
 equivalent to `chat_completions`, so every existing config keeps working
 byte-for-byte — the projected flat env is unchanged, and the default value
 is never emitted into the legacy `AZURE_OPENAI_POOL_*` override. Set
-`apiType` only when an endpoint speaks a non-default protocol.
+`wireApi` only when an endpoint speaks a non-default protocol.
 
 ```yaml
 azureOpenAI:
   deployments:
-    gpt_4_o: # no apiType ⇒ chat_completions (unchanged)
+    gpt_4_o: # no wireApi ⇒ chat_completions (unchanged)
       endpoints:
         - endpoint: https://my-resource.openai.azure.com/.../chat/completions?api-version=...
     gpt_5_codex:
       endpoints:
         - endpoint: https://api.openai.com/v1/responses
           region: eastus
-          apiType: openai_responses
+          wireApi: openai_responses
     claude_sonnet:
       endpoints:
         - endpoint: https://api.anthropic.com/v1/messages
           region: eastus
-          apiType: anthropic_messages
+          wireApi: anthropic_messages
 ```
 
-**Precedence.** `apiType` is a per-**endpoint** attribute, resolved on the
+**Precedence.** `wireApi` is a per-**endpoint** attribute, resolved on the
 same YAML → flat-env → typed-`Config` path as `capacity`/`priority`/`tpm`.
 In the flat-env representation it rides the `AZURE_OPENAI_POOL_<DEPLOYMENT>`
-override list (`[{suffix:...,apiType:openai_responses}]`), so a
+override list (`[{suffix:...,wireApi:openai_responses}]`), so a
 `process.env` override of that key wins over YAML, matching the rest of the
-merge precedence above. `apiType` never participates in routing — the pool
+merge precedence above. `wireApi` never participates in routing — the pool
 still picks _which_ endpoint to call by priority/capacity/cooldown; the
-resolved `apiType` then selects the `ProviderAdapter`
-([packages/aiclient/src/providerAdapter.ts](../aiclient/src/providerAdapter.ts))
+resolved `wireApi` then selects the `ProviderAdapter`
+([packages/aiclient/src/providers/](../aiclient/src/providerAdapter.ts))
 that encodes the request and decodes the response.
 
 This mirrors the shipping multi-provider pattern used elsewhere: opencode's
