+++ ts/packages/config/src/runtime/build.ts
@@ -30,12 +30,12 @@ import {
 import {
     AuthMode,
     authModeFromString,
-    apiTypeFromString,
+    wireApiFromString,
     Config,
     Deployment,
     DeploymentEndpoint,
     DeploymentMode,
-    type ApiType,
+    type WireApi,
     type AzureOpenAIConfig,
     type EmbeddingConfig,
     type EmbeddingProviderMode,
@@ -102,7 +102,7 @@ function makeEndpoint(
     capacity?: number,
     priority?: number,
     tpm?: number,
-    apiType?: ApiType,
+    wireApi?: WireApi,
 ): DeploymentEndpoint {
     const ep: DeploymentEndpoint = {
         endpoint,
@@ -112,7 +112,7 @@ function makeEndpoint(
         priority: priority ?? defaultPriority(mode),
         ...(capacity !== undefined ? { capacity } : {}),
         ...(tpm !== undefined ? { tpm } : {}),
-        ...(apiType !== undefined ? { apiType } : {}),
+        ...(wireApi !== undefined ? { wireApi } : {}),
     };
     return ep;
 }
@@ -330,7 +330,7 @@ interface PartialEndpoint {
     capacity?: number;
     priority?: number;
     tpm?: number;
-    apiType?: ApiType;
+    wireApi?: WireApi;
 }
 
 /** Per-suffix overrides parsed from `AZURE_OPENAI_POOL_<DEPLOYMENT>` JSON. */
@@ -341,7 +341,7 @@ interface PoolOverride {
     capacity?: number;
     priority?: number;
     tpm?: number;
-    apiType?: ApiType;
+    wireApi?: WireApi;
 }
 
 /**
@@ -384,9 +384,9 @@ function parsePoolOverride(
             if (typeof o.capacity === "number") ov.capacity = o.capacity;
             if (typeof o.priority === "number") ov.priority = o.priority;
             if (typeof o.tpm === "number") ov.tpm = o.tpm;
-            if (typeof o.apiType === "string") {
-                const at = apiTypeFromString(o.apiType);
-                if (at !== undefined) ov.apiType = at;
+            if (typeof o.wireApi === "string") {
+                const at = wireApiFromString(o.wireApi);
+                if (at !== undefined) ov.wireApi = at;
             }
             out.push(ov);
         }
@@ -468,7 +468,7 @@ function collectDeployments(
                 if (ov.capacity !== undefined) partial.capacity = ov.capacity;
                 if (ov.priority !== undefined) partial.priority = ov.priority;
                 if (ov.tpm !== undefined) partial.tpm = ov.tpm;
-                if (ov.apiType !== undefined) partial.apiType = ov.apiType;
+                if (ov.wireApi !== undefined) partial.wireApi = ov.wireApi;
             }
         }
     }
@@ -519,7 +519,7 @@ function collectDeployments(
                     partial.capacity,
                     partial.priority,
                     partial.tpm,
-                    partial.apiType,
+                    partial.wireApi,
                 ),
             );
         }
